feature/requirements.txt
fastapi
pydantic
python-jose[cryptography]
passlib[bcrypt]
uvicorn